requirements.txt
flet>=0.21.0
Pillow>=10.1.0
numpy>=1.26.0
opencv-python>=4.8.0